requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiosignal==1.3.1
aiosqlite==0.19.0
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
certifi==2023.11.17
chardet==3.0.4
charset-normalizer==3.3.2
configparser==6.0.0
contourpy==1.2.0
cycler==0.12.1
fonttools==4.47.0
frozenlist==1.4.0
greenlet==3.0.2
h11==0.8.1
h2==3.2.0
hpack==3.0.0
http3==0.6.7
hyperframe==5.2.0
idna==2.10
kiwisolver==1.4.5
magic-filter==1.0.12
marshmallow==3.20.1
matplotlib==3.8.2
multidict==6.0.4
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
pydantic==2.5.2
pydantic_core==2.14.5
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.9.0
urllib3==2.1.0
yarl==1.9.4
